Create liger object

This function allows creating liger object from multiple datasets of various forms (See rawData).

createLiger(
  rawData,
  modal = NULL,
  cellMeta = NULL,
  removeMissing = TRUE,
  addPrefix = TRUE,
  formatType = "10X",
  dataName = NULL,
  indicesName = NULL,
  indptrName = NULL,
  genesName = NULL,
  barcodesName = NULL,
  verbose = getOption("ligerVerbose"),
  ...,
  remove.missing = removeMissing,
  format.type = formatType,
  data.name = dataName,
  indices.name = indicesName,
  indptr.name = indptrName,
  genes.name = genesName,
  barcodes.name = barcodesName
)

Arguments

rawData

Named list of datasets. Required. Elements allowed include a matrix, a Seurat object, a SingleCellExperiment object, an AnnData object, a ligerDataset object or a filename to an HDF5 file. See detail for HDF5 reading.

modal

Character vector for modality setting. Currently options of "default", "rna", and "atac" are supported.

cellMeta

data.frame of metadata at single-cell level. Default NULL.

removeMissing

Logical. Whether to remove cells that do not have any counts and features not expressed in any cells from each dataset. Default TRUE. H5 based dataset with less than 8000 cells will be subset into memory.

addPrefix

Logical. Whether to add "<dataset name>_" as a prefix of cell identifiers (e.g. barcodes) to avoid duplicates in multiple libraries ( common with 10X data). Default TRUE

formatType

Select preset of H5 file structure. Current available options are "10X" and "AnnData". Can be either a single specification for all datasets or a character vector that match with each dataset.

dataName, indicesName, indptrName

The path in a H5 file for the raw sparse matrix data. These three types of data stands for the x, i, and p slots of a dgCMatrix-class object. Default NULL uses formatType preset.

genesName, barcodesName

The path in a H5 file for the gene names and cell barcodes. Default NULL uses formatType preset.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

...

Additional slot values that should be directly placed in object.

remove.missing, format.type, data.name, indices.name, indptr.name, genes.name, barcodes.name

Deprecated. See Usage section for replacement.

See also