Create in-memory ligerDataset object

Create in-memory ligerDataset object

createLigerDataset(
  rawData = NULL,
  modal = c("default", "rna", "atac"),
  normData = NULL,
  scaleData = NULL,
  featureMeta = NULL,
  ...
)

Arguments

rawData, normData, scaleData

A dgCMatrix-class object for the raw or normalized expression count or a dense matrix of scaled variable gene expression, respectively. Default NULL for all three but at lease one has to be specified.

modal

Name of modality for this dataset. Currently options of "default", "rna", and "atac" are supported. Default "default".

featureMeta

Data frame of feature metadata. Default NULL.

...

Additional slot data. See ligerDataset for detail. Given values will be directly placed at corresponding slots.

See also