Generate dot plot from input matrix with ComplexHeatmap

Generate dot plot from input matrix with ComplexHeatmap

.complexHeatmapDotPlot(
  colorMat,
  sizeMat,
  featureAnnDF = NULL,
  cellSplitVar = NULL,
  cellLabels = NULL,
  maxDotSize = 4,
  clusterFeature = FALSE,
  clusterCell = FALSE,
  legendColorTitle = "Matrix Value",
  legendSizeTitle = "Fraction Value",
  transpose = FALSE,
  baseSize = 8,
  cellTextSize = NULL,
  featureTextSize = NULL,
  cellTitleSize = NULL,
  featureTitleSize = NULL,
  legendTextSize = NULL,
  legendTitleSize = NULL,
  featureGrpRot = 0,
  viridisOption = "C",
  viridisDirection = -1,
  ...
)

Arguments

colorMat, sizeMat

Matrix of the same size. Values in colorMat will be visualized with color while values in sizeMat will be reflected by dot size.

featureAnnDF

Data frame of features containing feature names and grouping labels.

cellSplitVar

Split the cell orientation (default columns) by this variable.

cellLabels

Label to be shown on cell orientation.

maxDotSize

The maximum dot size. Default 4.

clusterFeature, clusterCell

Whether the feature/cell orientation (default rows/column, respectively) should be clustered. Default FALSE.

legendColorTitle, legendSizeTitle

The title for color bar and dot size legends, repectively. Default see "Matrix Value" and "Fraction Value".

transpose

Logical, whether to rotate the dot plot orientation. i.e. rows as cell aggregation and columns as features. Default FALSE.

baseSize

One-parameter control of all text sizes. Individual text element sizes can be controlled by other size arguments. "Title" sizes are 2 points larger than "text" sizes when being controlled by this. Default 8.

cellTextSize, featureTextSize, legendTextSize

Size of cell labels, feature label and legend text. Default NULL controls by baseSize.

cellTitleSize, featureTitleSize, legendTitleSize

Size of titles on cell and feature orientation and legend title. Default NULL controls by baseSize + 2.

featureGrpRot

Number of degree to rotate the feature grouping label. Default 0.

viridisOption, viridisDirection

See argument option and direction of viridis. Default "A" and -1.

...

Additional arguments passed to Heatmap.

Value

A HeatmapList object.