Find shared and dataset-specific markers

Applies various filters to genes on the shared (\(W\)) and dataset-specific (\(V\)) components of the factorization, before selecting those which load most significantly on each factor (in a shared or dataset-specific way).

getFactorMarkers(
  object,
  dataset1,
  dataset2,
  factorShareThresh = 10,
  datasetSpecificity = NULL,
  logFCThresh = 1,
  pvalThresh = 0.05,
  nGenes = 30,
  printGenes = FALSE,
  verbose = getOption("ligerVerbose"),
  factor.share.thresh = factorShareThresh,
  dataset.specificity = datasetSpecificity,
  log.fc.thresh = logFCThresh,
  pval.thresh = pvalThresh,
  num.genes = nGenes,
  print.genes = printGenes
)

Arguments

object

liger object with factorization results.

dataset1

Name of first dataset. Required.

dataset2

Name of second dataset. Required

factorShareThresh

Numeric. Only factors with a dataset specificity less than or equal to this threshold will be used. Default 10.

datasetSpecificity

Numeric vector. Pre-calculated dataset specificity if available. Length should match number of all factors available. Default NULL automatically calculates with calcDatasetSpecificity.

logFCThresh

Numeric. Lower log-fold change threshold for differential expression in markers. Default 1.

pvalThresh

Numeric. Upper p-value threshold for Wilcoxon rank test for gene expression. Default 0.05.

nGenes

Integer. Max number of genes to report for each dataset. Default 30.

printGenes

Logical. Whether to print ordered markers passing logFC, UMI and frac thresholds, when verbose = TRUE. Default FALSE.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

factor.share.thresh, dataset.specificity, log.fc.thresh, pval.thresh, num.genes, print.genes

Deprecated. See Usage section for replacement.

Value

A list object consisting of the following entries:

[value of `dataset1`]

data.frame of dataset1-specific markers

shared

data.frame of shared markers

[value of `dataset1`]

data.frame of dataset2-specific markers

num_factors_V1

A frequency table indicating the number of factors each marker appears, in dataset1

num_factors_V2

A frequency table indicating the number of factors each marker appears, in dataset2

Examples

library(dplyr)
result <- getFactorMarkers(pbmcPlot, dataset1 = "ctrl", dataset2 = "stim")
#> Sun Nov  5 17:13:10 2023 ... Performing wilcoxon test between datasets "ctrl" and "stim", 
#>                              basing on factor loading.
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%
print(class(result))
#> [1] "list"
print(names(result))
#> [1] "ctrl"           "shared"         "stim"           "num_factors_V1"
#> [5] "num_factors_V2"
result$shared %>% group_by(factor_num) %>% top_n(2, logFC)
#> # A tibble: 4 × 4
#> # Groups:   factor_num [2]
#>   feature factor_num logFC     pval
#>   <chr>        <int> <dbl>    <dbl>
#> 1 GPX1             1  3.05 8.08e-50
#> 2 RPS8             1  2.99 3.39e- 4
#> 3 CXCR4            2  2.09 8.41e- 4
#> 4 S100A6           2  3.40 1.52e- 6