Impute the query cell expression matrix

Impute query features from a reference dataset using KNN.

imputeKNN(
  object,
  reference,
  queries = NULL,
  nNeighbors = 20,
  weight = TRUE,
  norm = TRUE,
  scale = FALSE,
  verbose = getOption("ligerVerbose"),
  ...,
  knn_k = nNeighbors
)

Arguments

object

liger object with aligned factor loading computed in advance.

reference

Name of a dataset containing values to impute into query dataset(s).

queries

Names of datasets to be augmented by imputation. Should not include reference. Default NULL uses all datasets except the reference.

nNeighbors

The maximum number of nearest neighbors to search. Default 20.

weight

Logical. Whether to use KNN distances as weight matrix. Default FALSE.

norm

Logical. Whether to normalize the imputed data. Default TRUE.

scale

Logical. Whether to scale but not center the imputed data. Default TRUE.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

...

Optional arguments to be passed to normalize when norm = TRUE.

knn_k

Deprecated. See Usage section for replacement.

Value

The input object where queried ligerDataset

objects in datasets slot are replaced. These datasets will all be converted to ligerATACDataset class with an additional slot rawPeak to store the imputed peak counts, and normPeak for normalized imputed peak counts if norm = TRUE.

Examples

bmmc <- normalize(bmmc)
#> Mon Dec 11 16:10:53 2023 ... Normalizing dataset: rna
#> Mon Dec 11 16:10:53 2023 ... Normalizing dataset: atac
bmmc <- selectGenes(bmmc, datasets.use = "rna")
#> Warning: The `datasets.use` argument of `selectGenes.liger()` is deprecated as of
#> rliger2 1.99.0.
#> ℹ Please use the `useDatasets` argument instead.
#> ℹ The deprecated feature was likely used in the rliger2 package.
#>   Please report the issue at <https://github.com/welch-lab/liger/issues>.
#> Mon Dec 11 16:10:53 2023 ... Selecting variable features for dataset: rna
#> Mon Dec 11 16:10:53 2023 ...... 83 features selected out of 172 shared features
#> Mon Dec 11 16:10:53 2023 ... Finally 83 shared variable features selected.
bmmc <- scaleNotCenter(bmmc)
#> Mon Dec 11 16:10:53 2023 ... Scaling dataset: rna
#> Mon Dec 11 16:10:53 2023 ... Scaling dataset: atac
bmmc <- runINMF(bmmc, k = 20)
bmmc <- quantileNorm(bmmc)
bmmc <- normalizePeak(bmmc)
#> Mon Dec 11 16:10:54 2023 ... Normalizing rawPeak counts in dataset: atac
bmmc <- imputeKNN(bmmc, reference = "atac", queries = "rna")
#> Mon Dec 11 16:10:54 2023 ... Imputing all the datasets exept the reference dataset
#>                              Reference dataset: atac
#>                              Query datasets: rna
#> Mon Dec 11 16:10:54 2023 ... Normalizing rawPeak counts in dataset: rna