Check if a liger or ligerDataset object is made of HDF5 file

Check if a liger or ligerDataset object is made of HDF5 file

isH5Liger(object, dataset = NULL)

Arguments

object

A liger or ligerDataset object.

dataset

If object is of liger class, check a specific dataset. If NULL, Check if all datasets are made of HDF5 file. Default NULL.

Value

TRUE or FALSE for the specified check.

Examples

isH5Liger(pbmc)
#> [1] FALSE
isH5Liger(pbmc, "ctrl")
#> [1] FALSE
ctrl <- dataset(pbmc, "ctrl")
isH5Liger(ctrl)
#> [1] FALSE