Find DEG between two groups

Find DEG between two groups. Two methods are supported: "wilcoxon" and "pseudoBulk". Wilcoxon rank sum test is performed on single-cell level, while pseudo-bulk method aggregates cells basing on biological replicates and calls bulk RNAseq DE methods, DESeq2 wald test. When real biological replicates are not available, pseudo replicates can be generated. Please see below for detailed scenario usage.

runPairwiseDEG(
  object,
  group1,
  group2,
  variable1 = NULL,
  variable2 = NULL,
  method = c("wilcoxon", "pseudoBulk"),
  usePeak = FALSE,
  useReplicate = NULL,
  nPsdRep = 5,
  seed = 1,
  verbose = getOption("ligerVerbose")
)

runMarkerDEG(
  object,
  conditionBy = NULL,
  splitBy = NULL,
  method = c("wilcoxon", "pseudoBulk"),
  useDatasets = NULL,
  usePeak = FALSE,
  useReplicate = NULL,
  nPsdRep = 5,
  seed = 1,
  verbose = getOption("ligerVerbose")
)

Arguments

object

A liger object, with normalized data available

group1, group2, variable1, variable2

Condition specification. See ?runPairwiseDEG section Pairwise DEG Scenarios for detail.

method

DEG test method to use. Choose from "wilcoxon" or "pseudoBulk". Default "wilcoxon"

usePeak

Logical. Whether to use peak count instead of gene count. Only supported when ATAC datasets are involved. Default FALSE.

useReplicate

cellMeta variable of biological replicate annotation. Only used with method = "pseudoBulk". Default NULL will create nPsdRep pseudo replicates per group.

nPsdRep

Number of pseudo replicates to create. Only used when method = "pseudoBulk", useReplicate = NULL. Default 5.

seed

Random seed to use for pseudo-replicate generation. Default 1.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

conditionBy

cellMeta variable(s). Marker detection will be performed for each level of this variable. Multiple variables will be combined. Default NULL uses default cluster.

splitBy

Split data by cellMeta variable(s) here and identify markers for conditionBy within each chunk. Default NULL.

useDatasets

Datasets to perform marker detection within. Default NULL will use all datasets.

Value

A data.frame with DEG information

Pairwise DEG Scenarios

Users can select classes of cells from a variable in cellMeta. variable1 and variable2 are used to specify a column in cellMeta, and group1 and group2 are used to specify existing classes from variable1 and variable2, respectively. When variable2 is missing, group2 will be considered from variable1.

For example, when variable1 = "celltype" and variable2 = NULL, group1 and group2 should be valid cell types in object$celltype.

When variable1 is "celltype" and variable2 is "gender", group1 should be a valid cell type from object$celltype and group2 should be a valid class from object$gender.

When both variable1 and variable2 are missing, group1 and group2 should be valid index of cells in object.

Marker Detection Scenarios

Marker detection is generally performed in a one vs. rest manner. The grouping of such condition is specified by conditionBy, which should be a column name in cellMeta. When splitBy is specified as another variable name in cellMeta, the marker detection will be iteratively done for each level of splitBy variable.

For example, when conditionBy = "celltype" and splitBy = NULL, marker detection will be performed by comparing all cells of "celltype_i" against all other cells, and etc.

When conditionBy = "celltype" and splitBy = "gender", marker detection will be performed by comparing "celltype_i" cells from "gender_i" against other cells from "gender_i", and etc.

Examples

# Compare between cluster "0" and cluster "1"
degStats <- runPairwiseDEG(pbmcPlot, group1 = 0, group2 = 1,
                           variable1 = "leiden_cluster")
# Compare between all cells from cluster "5" and
# all cells from dataset "stim"
degStats <- runPairwiseDEG(pbmcPlot, group1 = "5", group2 = "stim",
                           variable1 = "leiden_cluster",
                           variable2 = "dataset")
# Identify markers for each cluster
markerStats <- runMarkerDEG(pbmcPlot, conditionBy = "leiden_cluster")
# Identify dataset markers within each cluster
markerStatsList <- runMarkerDEG(pbmcPlot, conditionBy = "dataset",
                                splitBy = "leiden_cluster")