ligerCommand object: Record the input and time of a LIGER function call

ligerCommand object: Record the input and time of a LIGER function call

# S4 method for ligerCommand
show(object)

Arguments

object

A ligerCommand object

Slots

funcName

Name of the function

time

A time stamp object

call

A character string converted from system call

parameters

List of all arguments except the liger object. Large object are summarized to short string.

objSummary

List of attributes of the liger object as a snapshot when command is operated.

ligerVersion

Character string converted from packageVersion("rliger2").

dependencyVersion

Named character vector of version number, if any dependency library has a chance to be included by the function. A dependency might only be invoked under certain conditions, such as using an alternative algorithm, which a call does not actually reach to, but it would still be included for this call.

Examples

pbmc <- normalize(pbmc)
#> Mon Dec 11 16:10:55 2023 ... Normalizing dataset: ctrl
#> Mon Dec 11 16:10:55 2023 ... Normalizing dataset: stim
cmd <- commands(pbmc, "normalize")
cmd
#> A liger command record, performed at 12-11-2023 16:10:55 EST 
#> Call: normalize.liger(pbmc) 
#> Parameters:
#>     useDatasets : "ctrl", "stim" 
#>     verbose : TRUE