ligerDataset class

Object for storing dastaset specific information. Will be embedded within a higher level liger object

# S4 method for ligerDataset
show(object)

# S4 method for ligerDataset
dim(x)

# S4 method for ligerDataset
dimnames(x)

# S4 method for ligerDataset,list
dimnames(x) <- value

# S4 method for ligerDataset,index,missing,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for ligerDataset,missing,index,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for ligerDataset,index,index,ANY
[(x, i, j, ..., drop = TRUE)

rawData(x, dataset = NULL)

rawData(x, dataset = NULL, check = TRUE) <- value

# S4 method for ligerDataset
rawData(x, dataset = NULL)

# S4 method for ligerDataset,ANY,ANY,matrixLike_OR_NULL
rawData(x, dataset = NULL, check = TRUE) <- value

# S4 method for ligerDataset,ANY,ANY,H5D
rawData(x, dataset = NULL, check = TRUE) <- value

normData(x, dataset = NULL)

normData(x, dataset = NULL, check = TRUE) <- value

# S4 method for ligerDataset
normData(x, dataset = NULL)

# S4 method for ligerDataset,ANY,ANY,matrixLike_OR_NULL
normData(x, dataset = NULL, check = TRUE) <- value

# S4 method for ligerDataset,ANY,ANY,H5D
normData(x, dataset = NULL, check = TRUE) <- value

scaleData(x, dataset = NULL)

scaleData(x, dataset = NULL, check = TRUE) <- value

# S4 method for ligerDataset,missing
scaleData(x, dataset = NULL)

# S4 method for ligerDataset,ANY,ANY,matrixLike_OR_NULL
scaleData(x, dataset = NULL, check = TRUE) <- value

# S4 method for ligerDataset,ANY,ANY,H5D
scaleData(x, dataset = NULL, check = TRUE) <- value

# S4 method for ligerDataset,ANY,ANY,H5Group
scaleData(x, dataset = NULL, check = TRUE) <- value

scaleUnsharedData(x, dataset = NULL)

scaleUnsharedData(x, check = TRUE) <- value

# S4 method for ligerDataset,missing
scaleUnsharedData(x, dataset = NULL)

# S4 method for ligerDataset,ANY,matrixLike_OR_NULL
scaleUnsharedData(x, check = TRUE) <- value

# S4 method for ligerDataset,ANY,H5D
scaleUnsharedData(x, check = TRUE) <- value

# S4 method for ligerDataset,ANY,H5Group
scaleUnsharedData(x, check = TRUE) <- value

getMatrix(x, slot = "rawData", dataset = NULL, returnList = FALSE)

# S4 method for ligerDataset,ANY,missing,missing
getMatrix(
  x,
  slot = c("rawData", "normData", "scaleData", "scaleUnsharedData", "H", "V", "U", "A",
    "B"),
  dataset = NULL
)

h5fileInfo(x, info = NULL)

h5fileInfo(x, info = NULL, check = TRUE) <- value

# S4 method for ligerDataset
h5fileInfo(x, info = NULL)

# S4 method for ligerDataset
h5fileInfo(x, info = NULL, check = TRUE) <- value

getH5File(x, dataset = NULL)

# S4 method for ligerDataset,missing
getH5File(x, dataset = NULL)

featureMeta(x, check = NULL)

featureMeta(x, check = TRUE) <- value

# S4 method for ligerDataset
featureMeta(x, check = NULL)

# S4 method for ligerDataset
featureMeta(x, check = TRUE) <- value

# S3 method for ligerDataset
cbind(x, ..., deparse.level = 1)

Arguments

x, object

A ligerDataset object.

value

See detail sections for requirements

i, j

Feature and cell index for `[` method. For `[[` method, use a single variable name with i and j is not applicable.

...

See detailed sections for explanation.

drop

Not applicable.

dataset

Not applicable for ligerDataset methods.

check

Whether to perform object validity check on setting new value.

slot

The slot name when using getMatrix.

returnList

Not applicable for ligerDataset methods.

info

Name of the entry in h5fileInfo slot.

deparse.level

Not used here.

Slots

rawData

Raw data.

normData

Normalized data

scaleData

Scaled data, usually with subset variable features

scaleUnsharedData

Scaled data of features not shared with other datasets

varUnsharedFeatures

Variable features not shared with other datasets

V

matrix

A

matrix

B

matrix

H

matrix

U

matrix

h5fileInfo

list

featureMeta

Feature metadata, DataFrame

colnames

character

rownames

character

Dimensionality

For a ligerDataset object, the column orientation is assigned for cells and rows are for features. Therefore, for ligerDataset objects, dim() returns a numeric vector of two numbers which are number of features and number of cells. dimnames() returns a list of two character vectors, which are the feature names and the cell barcodes.

For direct call of dimnames<- method, value should be a list with a character vector of feature names as the first element and cell identifiers as the second element. For colnames<- method, the character vector of cell identifiers. For rownames<- method, the character vector of feature names.

Subsetting

For more detail of subsetting a liger object or a ligerDataset object, please check out subsetLiger and subsetLigerDataset. Here, we set the S4 method "single-bracket" [ as a quick wrapper to subset a ligerDataset object. i and j serves as feature and cell subscriptor, respectively, which can be any valid index refering the available features and cells in a dataset. ... arugments are passed to subsetLigerDataset so that advanced options are allowed.

Matrix access

For ligerDataset object, rawData(), normData, scaleData() and scaleUnsharedData() methods are exported for users to access the corresponding feature expression matrix. Replacement methods are also available to modify the slots.

For other matrices, such as the \(H\) and \(V\), which are dataset specific, please use getMatrix() method with specifying slot name. Directly accessing slot with @ is generally not recommended.

H5 file and information access

A ligerDataset object has a slot called h5fileInfo, which is a list object. The first element is called $H5File, which is an H5File class object and is the connection to the input file. The second element is $filename which stores the absolute path of the H5 file in the current machine. The third element $formatType stores the name of preset being used, if applicable. The other following keys pair with paths in the H5 file that point to specific data for constructing a feature expression matrix.

h5fileInfo() method access the list described above and simply retrieves the corresponding value. When info = NULL, returns the whole list. When length(info) == 1, returns the requested list value. When more info requested, returns a subset list.

The replacement method modifies the list elements and corresponding slot value (if applicable) at the same time. For example, running h5fileInfo(obj, "rawData") <- newPath not only updates the list, but also updates the rawData slot with the H5D class data at "newPath" in the H5File object.

getH5File() is a wrapper and is equivalent to h5fileInfo(obj, "H5File").

Feature metadata access

A slot featureMeta is included for each ligerDataset object. This slot requires a DataFrame-class object, which is the same as cellMeta slot of a liger object. However, the associated S4 methods only include access to the whole table for now. Internal information access follows the same way as data.frame operation. For example, featureMeta(ligerD)$nCell or featureMeta(ligerD)[varFeatures(ligerObj), "gene_var"].

Concatenate ligerDataset

cbind() method is implemented for concatenating ligerDataset objects by cells. When applying, all feature expression matrix will be merged with taking a union of all features for the rows.

Examples

ctrl <- dataset(pbmc, "ctrl")

# Methods for base generics
ctrl
#> An object of class ligerDataset with 300 cells
#> rawData: 266 features
print(ctrl)
#> An object of class ligerDataset with 300 cells
#> rawData: 266 features
dim(ctrl)
#> [1] 266 300
ncol(ctrl)
#> [1] 300
nrow(ctrl)
#> [1] 266
colnames(ctrl)[1:5]
#> [1] "ctrl_AAACATACCTCGCT.1" "ctrl_AAACGGCTCTTCGC.1" "ctrl_AACACTCTAAGTAG.1"
#> [4] "ctrl_AACCGCCTCAGGAG.1" "ctrl_AACGTTCTTCCGTC.1"
rownames(ctrl)[1:5]
#> [1] "ISG15"    "ID3"      "RPL11"    "MARCKSL1" "RPS8"    
ctrl[1:5, 1:5]
#> An object of class ligerDataset with 5 cells
#> rawData: 5 features

# rliger generics
## raw data
m <- rawData(ctrl)
class(m)
#> [1] "dgCMatrix"
#> attr(,"package")
#> [1] "Matrix"
dim(m)
#> [1] 266 300
## normalized data
pbmc <- normalize(pbmc)
#> Mon Dec 11 16:10:55 2023 ... Normalizing dataset: ctrl
#> Mon Dec 11 16:10:55 2023 ... Normalizing dataset: stim
ctrl <- dataset(pbmc, "ctrl")
m <- normData(ctrl)
class(m)
#> [1] "dgCMatrix"
#> attr(,"package")
#> [1] "Matrix"
dim(m)
#> [1] 266 300
## scaled data
pbmc <- selectGenes(pbmc)
#> Mon Dec 11 16:10:55 2023 ... Selecting variable features for dataset: ctrl
#> Mon Dec 11 16:10:55 2023 ...... 168 features selected out of 249 shared features
#> Mon Dec 11 16:10:55 2023 ... Selecting variable features for dataset: stim
#> Mon Dec 11 16:10:55 2023 ...... 166 features selected out of 249 shared features
#> Mon Dec 11 16:10:55 2023 ... Finally 173 shared variable features selected.
pbmc <- scaleNotCenter(pbmc)
#> Mon Dec 11 16:10:56 2023 ... Scaling dataset: ctrl
#> Mon Dec 11 16:10:56 2023 ... Scaling dataset: stim
ctrl <- dataset(pbmc, "ctrl")
m <- scaleData(ctrl)
class(m)
#> [1] "dgCMatrix"
#> attr(,"package")
#> [1] "Matrix"
dim(m)
#> [1] 173 300
n <- scaleData(pbmc, "ctrl")
identical(m, n)
#> [1] TRUE
## Any other matrices
pbmc <- online_iNMF(pbmc, k = 20, miniBatch_size = 100)
#> Warning: `online_iNMF()` was deprecated in rliger2 1.99.0.
#> ℹ Please use the `method` argument of `runIntegration()` instead.
ctrl <- dataset(pbmc, "ctrl")
V <- getMatrix(ctrl, "V")
V[1:5, 1:5]
#>              Factor_1     Factor_2     Factor_3     Factor_4 Factor_5
#> ISG15    1.000000e-16 1.000000e-16 1.000000e-16 7.374195e-04    1e-16
#> ID3      1.000000e-16 3.322680e-03 5.289525e-03 1.000000e-16    1e-16
#> RPL11    1.000000e-16 8.240158e-03 1.000000e-16 4.101785e-03    1e-16
#> MARCKSL1 1.000000e-16 1.000000e-16 1.000000e-16 1.000000e-16    1e-16
#> RPS8     5.607769e-03 1.000000e-16 1.000000e-16 7.515360e-05    1e-16
Vs <- getMatrix(pbmc, "V")
length(Vs)
#> [1] 2
names(Vs)
#> [1] "ctrl" "stim"
identical(Vs$ctrl, V)
#> [1] TRUE