Linking genes to putative regulatory elements

Evaluate the relationships between pairs of genes and peaks based on specified distance metric. Usually used for inferring the correlation between gene expression and imputed peak counts for datasets without the modality originally (i.e. applied to imputeKNN result).

linkGenesAndPeaks(
  object,
  useDataset,
  pathToCoords,
  useGenes = NULL,
  method = c("spearman", "pearson", "kendall"),
  alpha = 0.05,
  verbose = getOption("ligerVerbose"),
  path_to_coords = pathToCoords,
  genes.list = useGenes,
  dist = method
)

Arguments

object

A liger object, with datasets that is of ligerATACDataset class in the datasets slot.

useDataset

Name of one dataset, with both normalized gene expression and normalized peak counts available.

pathToCoords

Path tothe gene coordinates file, usually a BED file.

useGenes

Character vector of gene names to be tested. Default NULL uses all genes available in useDataset.

method

Choose the type of correlation to calculate, from "spearman", "pearson" and "kendall". Default "spearman"

alpha

Numeric, significance threshold for correlation p-value. Peak-gene correlations with p-values below this threshold are considered significant. Default 0.05.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

path_to_coords, genes.list, dist

Deprecated. See Usage section for replacement.

Value

A sparse matrix with peak names as rows and gene names as columns, with each element indicating the correlation between peak i and gene j, 0 if the gene and peak are not significantly linked.

See also

Examples

bmmc <- normalize(bmmc)
#> Sun Nov  5 11:37:42 2023 ... Normalizing dataset: rna
#> Sun Nov  5 11:37:42 2023 ... Normalizing dataset: atac
bmmc <- selectGenes(bmmc)
#> Sun Nov  5 11:37:42 2023 ... Selecting variable features for dataset: rna
#> Sun Nov  5 11:37:42 2023 ...... 83 features selected out of 172 shared features
#> Sun Nov  5 11:37:42 2023 ... Selecting variable features for dataset: atac
#> Sun Nov  5 11:37:42 2023 ...... 126 features selected out of 172 shared features
#> Sun Nov  5 11:37:42 2023 ... Finally 135 shared variable features selected.
bmmc <- scaleNotCenter(bmmc)
#> Sun Nov  5 11:37:42 2023 ... Scaling dataset: rna
#> Sun Nov  5 11:37:42 2023 ... Scaling dataset: atac
bmmc <- online_iNMF(bmmc, miniBatch_size = 80)
bmmc <- quantileNorm(bmmc)
bmmc <- normalizePeak(bmmc)
#> Sun Nov  5 11:37:42 2023 ... Normalizing rawPeak counts in dataset: atac
bmmc <- imputeKNN(bmmc, reference = "atac", queries = "rna")
#> Sun Nov  5 11:37:42 2023 ... Imputing all the datasets exept the reference dataset
#>                              Reference dataset: atac
#>                              Query datasets: rna
#> Sun Nov  5 11:37:42 2023 ... Normalizing rawPeak counts in dataset: rna
corr <- linkGenesAndPeaks(
    bmmc, useDataset = "rna",
    pathToCoords = system.file("extdata/hg19_genes.bed", package = "rliger2")
)
#> Sun Nov  5 11:37:42 2023 ... 172 genes to be tested against 995 peaks
#> Sun Nov  5 11:37:42 2023 ... Calculating correlation for gene-peak pairs...
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   8%  |                                                                              |======                                                                |   8%  |                                                                              |=======                                                               |   9%  |                                                                              |========                                                              |  11%  |                                                                              |==========                                                            |  14%  |                                                                              |===========                                                           |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  26%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |========================                                              |  34%  |                                                                              |=========================                                             |  36%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  40%  |                                                                              |============================                                          |  41%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |=================================                                     |  48%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |=====================================                                 |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |========================================                              |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  62%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |==================================================                    |  72%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |======================================================                |  77%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  85%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  89%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |====================================================================  |  98%  |                                                                              |===================================================================== |  98%  |                                                                              |======================================================================| 100%