Perform online iNMF on scaled datasets

Perform online integrative non-negative matrix factorization to represent multiple single-cell datasets in terms of \(H\), \(W\), and \(V\) matrices. It optimizes the iNMF objective function using online learning (non-negative least squares for H matrix, hierarchical alternating least squares for \(W\) and \(V\) matrices), where the number of factors is set by k. The function allows online learning in 3 scenarios:

Fully observed datasets;

Iterative refinement using continually arriving datasets;

Projection of new datasets without updating the existing factorization

All three scenarios require fixed memory independent of the number of cells.

For each dataset, this factorization produces an \(H\) matrix (k by cell), a \(V\) matrix (genes by k), and a shared \(W\) matrix (genes by k). The \(H\) matrices represent the cell factor loadings. \(W\) is identical among all datasets, as it represents the shared components of the metagenes across datasets. The \(V\) matrices represent the dataset-specific components of the metagenes.

online_iNMF(
  object,
  X_new = NULL,
  projection = FALSE,
  W.init = NULL,
  V.init = NULL,
  A.init = NULL,
  B.init = NULL,
  k = 20,
  lambda = 5,
  max.epochs = 5,
  miniBatch_max_iters = 1,
  miniBatch_size = 5000,
  h5_chunk_size = 1000,
  seed = 123,
  verbose = getOption("ligerVerbose")
)

Arguments

object

liger object. Scaled data required.

X_new

New datasets for scenario 2 or scenario 3. See detail for usage.

projection

Perform data integration with scenario 3. See description. Default FALSE.

W.init

Optional initialization for W. See detail. Default NULL.

V.init

Optional initialization for V. See detail. Default NULL.

A.init

Optional initialization for A. See detail. Default NULL.

B.init

Optional initialization for B. See detail. Default NULL.

k

Inner dimension of factorization--number of metagenes. A value in the range 20-50 works well for most analyses. Default 20.

lambda

Regularization parameter. Larger values penalize dataset-specific effects more strongly (i.e. alignment should increase as lambda increases). We recommend always using the default value except possibly for analyses with relatively small differences (biological replicates, male/female comparisons, etc.) in which case a lower value such as 1.0 may improve reconstruction quality. Default 5.0.

max.epochs

The number of epochs to iterate through. See detail. Default 5.

miniBatch_max_iters

Maximum number of block coordinate descent (HALS algorithm) iterations to perform for each update of \(W\) and \(V\). Default 1. Changing this parameter is not recommended.

miniBatch_size

Total number of cells in each minibatch. See detail. Default 5000.

h5_chunk_size

Chunk size of input hdf5 files. Default 1000. The chunk size should be no larger than the batch size.

seed

Random seed to allow reproducible results. Default 123.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

Value

object with W slot updated with resulting \(W\)

matrix; the H, V, A and B slots of each ligerDataset object in datasets slot is updated with the corresponding result matrices.

Details

For optional initialization, W.init must be a matrix object with number of rows equal to number of variable genes (denoted as g) and number of columns equal to k. Any of V.init, A and B must be a list object of n matrices where n is the number of datasets in object. For V.init, each matrix should be of size g x k. For A.init, each matrix should be k x k and for B.init, each matrix should be g x k.

Minibatch iterations is performed on small subset of cells. The exact minibatch size applied on each dataset is miniBatch_size multiplied by the proportion of cells in this dataset out of all cells. The setting of miniBatch_size is by default 5000, which is reasonable. However, a smaller value such as 1000 may be necessary for analyzing very small datasets. In general, miniBatch_size should be no larger than the number of cells in the smallest dataset. An epoch is one completion of calculation on all cells after a number of iterations of minibatches. Therefore, the total number of iterations is determined by the setting of max.epochs, total number of cells, and miniBatch_size.

Examples

pbmc <- normalize(pbmc)
#> Mon Apr 24 14:49:35 2023 ... Normalizing dataset: ctrl
#> Mon Apr 24 14:49:35 2023 ... Normalizing dataset: stim
pbmc <- selectGenes(pbmc)
#> Mon Apr 24 14:49:35 2023 ... Selecting HVG for dataset: ctrl
#> Mon Apr 24 14:49:35 2023 ...   168 shared variable features selected
#> Mon Apr 24 14:49:35 2023 ... Selecting HVG for dataset: stim
#> Mon Apr 24 14:49:35 2023 ...   166 shared variable features selected
#> Mon Apr 24 14:49:35 2023 ... Finally 173 shared variable features selected.
pbmc <- scaleNotCenter(pbmc)
#> Mon Apr 24 14:49:35 2023 ... Scaling dataset: ctrl
#> Mon Apr 24 14:49:35 2023 ... Scaling dataset: stim
# Minibatch size has to be less than number of cell in the smallest dataset
# Scenario 1
pbmc <- online_iNMF(pbmc, miniBatch_size = 100)
#> Mon Apr 24 14:49:35 2023 ... 2 new datasets detected.
#> Mon Apr 24 14:49:35 2023 ... Scenario 1, init arguments ignored.
#> Mon Apr 24 14:49:35 2023 ... Initialization done.
#> Mon Apr 24 14:49:35 2023 ... Starting Online iNMF...
#>   |                                                                              |==                                                                    |   3%  |                                                                              |=====                                                                 |   7%  |                                                                              |=======                                                               |  10%  |                                                                              |=========                                                             |  13%  |                                                                              |============                                                          |  17%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  23%  |                                                                              |===================                                                   |  27%  |                                                                              |=====================                                                 |  30%  |                                                                              |=======================                                               |  33%  |                                                                              |==========================                                            |  37%  |                                                                              |============================                                          |  40%  |                                                                              |==============================                                        |  43%  |                                                                              |=================================                                     |  47%  |                                                                              |===================================                                   |  50%  |                                                                              |=====================================                                 |  53%  |                                                                              |========================================                              |  57%  |                                                                              |==========================================                            |  60%  |                                                                              |============================================                          |  63%  |                                                                              |===============================================                       |  67%  |                                                                              |=================================================                     |  70%  |                                                                              |===================================================                   |  73%  |                                                                              |======================================================                |  77%  |                                                                              |========================================================              |  80%  |                                                                              |==========================================================            |  83%  |                                                                              |=============================================================         |  87%  |                                                                              |===============================================================       |  90%  |                                                                              |=================================================================     |  93%  |                                                                              |====================================================================  |  97%  |                                                                              |======================================================================| 100%
#> Mon Apr 24 14:49:35 2023 ... Calculating metagene loadings...
# Scenario 2
# Fake new dataset by increasing all non-zero value in "ctrl" by 1
ctrl2 <- rawData(dataset(pbmc, "ctrl"))
ctrl2@x <- ctrl2@x + 1
colnames(ctrl2) <- paste0(colnames(ctrl2), 2)
pbmc2 <- online_iNMF(pbmc, k = 20, X_new = list(ctrl2 = ctrl2),
                     miniBatch_size = 100)
#> Checking in memory ligerDataset validity
#> Checking liger object validity
#> Mon Apr 24 14:49:35 2023 ... Normalizing dataset: ctrl2
#> Mon Apr 24 14:49:35 2023 ... Scaling dataset: ctrl2
#> Mon Apr 24 14:49:35 2023 ... 1 new datasets detected.
#> Mon Apr 24 14:49:35 2023 ... Scenario 2, initiating parameters
#> Mon Apr 24 14:49:35 2023 ... Initialization done.
#> Mon Apr 24 14:49:35 2023 ... Starting Online iNMF...
#>   |                                                                              |=====                                                                 |   7%  |                                                                              |=========                                                             |  13%  |                                                                              |==============                                                        |  20%  |                                                                              |===================                                                   |  27%  |                                                                              |=======================                                               |  33%  |                                                                              |============================                                          |  40%  |                                                                              |=================================                                     |  47%  |                                                                              |=====================================                                 |  53%  |                                                                              |==========================================                            |  60%  |                                                                              |===============================================                       |  67%  |                                                                              |===================================================                   |  73%  |                                                                              |========================================================              |  80%  |                                                                              |=============================================================         |  87%  |                                                                              |=================================================================     |  93%  |                                                                              |======================================================================| 100%
#> Mon Apr 24 14:49:35 2023 ... Calculating metagene loadings...
# Scenario 3
pbmc3 <- online_iNMF(pbmc, k = 20, X_new = list(ctrl2 = ctrl2),
                     miniBatch_size = 100, projection = TRUE)
#> Checking in memory ligerDataset validity
#> Checking liger object validity
#> Mon Apr 24 14:49:35 2023 ... Normalizing dataset: ctrl2
#> Mon Apr 24 14:49:35 2023 ... Scaling dataset: ctrl2
#> Mon Apr 24 14:49:35 2023 ... 1 new datasets detected.
#> Mon Apr 24 14:49:35 2023 ... Scenario 3, metagene projection