Perform iNMF on scaled datasets

Performs integrative non-negative matrix (iNMF) factorization to return factorized \(H\), \(W\), and \(V\) matrices. It optimizes the iNMF objective function using block coordinate descent (alternating non-negative least squares), where the number of factors is set by k. TODO: include objective function equation here in documentation (using deqn)

For each dataset, this factorization produces an \(H\) matrix (cells by k), a \(V\) matrix (k by genes), and a shared \(W\) matrix (k by genes). The \(H\) matrices represent the cell factor loadings. \(W\) is held consistent among all datasets, as it represents the shared components of the metagenes across datasets. The \(V\) matrices represent the dataset-specific components of the metagenes.

optimizeALS(
  object,
  k,
  lambda = 5,
  thresh = 1e-06,
  maxIter = 30,
  nrep = 1,
  H.init = NULL,
  W.init = NULL,
  V.init = NULL,
  useUnshared = FALSE,
  seed = 1,
  readH5 = "auto",
  verbose = getOption("ligerVerbose"),
  max.iters = maxIter,
  use.unshared = useUnshared,
  rand.seed = seed,
  print.obj = NULL
)

# S4 method for liger
optimizeALS(
  object,
  k,
  lambda = 5,
  thresh = 1e-06,
  maxIter = 30,
  nrep = 1,
  H.init = NULL,
  W.init = NULL,
  V.init = NULL,
  useUnshared = FALSE,
  seed = 1,
  readH5 = "auto",
  verbose = getOption("ligerVerbose"),
  max.iters = maxIter,
  use.unshared = useUnshared,
  rand.seed = seed,
  print.obj = NULL
)

# S4 method for list
optimizeALS(
  object,
  k,
  lambda = 5,
  thresh = 1e-06,
  maxIter = 30,
  nrep = 1,
  H.init = NULL,
  W.init = NULL,
  V.init = NULL,
  useUnshared = FALSE,
  seed = 1,
  readH5 = "auto",
  verbose = getOption("ligerVerbose"),
  max.iters = maxIter,
  use.unshared = useUnshared,
  rand.seed = seed,
  print.obj = NULL
)

Arguments

object

A liger object or a named list of matrix object, where the names represents dataset names and matrices are scaled on the same set of variable features, with rows as features and columns as cells.

k

Inner dimension of factorization (number of factors). Run suggestK to determine appropriate value; a general rule of thumb is that a higher k will be needed for datasets with more sub-structure.

lambda

Regularization parameter. Larger values penalize dataset-specific effects more strongly (i.e. alignment should increase as lambda increases). Default 5.

thresh

Convergence threshold. Convergence occurs when \(|obj_0-obj|/(mean(obj_0,obj)) < thresh\). Default 1e-6.

maxIter

Maximum number of block coordinate descent iterations to perform. Default 30.

nrep

Number of restarts to perform (iNMF objective function is non-convex, so taking the best objective from multiple successive initialization is recommended). For easier reproducibility, this increments the random seed by 1 for each consecutive restart, so future factorization of the same dataset can be run with one rep if necessary. Default 1.

H.init

Initial values to use for \(H\) matrices. A list object where each element is the initial \(H\) matrix of each dataset. Default NULL.

W.init

Initial values to use for \(W\) matrix. A matrix object. Default NULL.

V.init

Initial values to use for \(V\) matrices. A list object where each element is the initial \(V\) matrix of each dataset. Default NULL.

useUnshared

Logical, whether to include unshared variable features and run optimizeUANLS algorithm. Defaul FALSE. Running selectGenes with unshared = TRUE and then running scaleNotCenter is required.

seed

Random seed to allow reproducible results. Default 1.

readH5

TRUE to force reading H5 based data into memory and conduct factorization. "auto" reads H5 dataset with less than 8000 cells. FALSE will stop users from running if H5 data presents.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

max.iters, use.unshared, rand.seed

Deprecated. See Usage section for replacement.

print.obj

Defunct. Whether to print objective function values after convergence when verbose = TRUE. Now always print when verbose.

Value

object with W slot updated with the result \(W\)

matrix, and the H and V slots of each ligerDataset object in the datasets slot updated with the dataset specific \(H\) and \(V\) matrix, respectively.

Examples

pbmc <- normalize(pbmc)
#> Mon Apr 24 14:49:35 2023 ... Normalizing dataset: ctrl
#> Mon Apr 24 14:49:35 2023 ... Normalizing dataset: stim
pbmc <- selectGenes(pbmc)
#> Mon Apr 24 14:49:35 2023 ... Selecting HVG for dataset: ctrl
#> Mon Apr 24 14:49:35 2023 ...   168 shared variable features selected
#> Mon Apr 24 14:49:35 2023 ... Selecting HVG for dataset: stim
#> Mon Apr 24 14:49:35 2023 ...   166 shared variable features selected
#> Mon Apr 24 14:49:35 2023 ... Finally 173 shared variable features selected.
pbmc <- scaleNotCenter(pbmc)
#> Mon Apr 24 14:49:35 2023 ... Scaling dataset: ctrl
#> Mon Apr 24 14:49:35 2023 ... Scaling dataset: stim
# Only running a few iterations for fast examples
pbmc <- optimizeALS(pbmc, k = 20, maxIter = 2)
#> Mon Apr 24 14:49:37 2023 ... Start iNMF with seed: 1...
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%
#> Mon Apr 24 14:49:40 2023 ... Finished in 3.6766140460968 secs, 2 iterations. 
#>                              Max iterations set: 2
#>                              Final objective delta: 0.278833145358608
#> Mon Apr 24 14:49:40 2023 ... Objective: 44307.2622916006
#> Mon Apr 24 14:49:40 2023 ... Best results with seed 1