Perform factorization for new lambda value

Uses an efficient strategy for updating that takes advantage of the information in the existing factorization; uses previous k. Recommended mainly when re-optimizing for higher lambda and when new lambda value is significantly different; otherwise may not return optimal results.

optimizeNewLambda(
  object,
  new.lambda,
  thresh = 1e-04,
  max.iters = 100,
  rand.seed = 1,
  verbose = getOption("ligerVerbose")
)

Arguments

object

liger object. Should call optimizeALS in advance.

new.lambda

Numeric regularization parameter. Larger values penalize dataset-specific effects more strongly.

thresh

Convergence threshold. Convergence occurs when \(|obj_0-obj|/(mean(obj_0,obj)) < thresh\). Default 1e-4.

max.iters

Maximum number of block coordinate descent iterations to perform. Default 100.

rand.seed

Random seed to allow reproducible results. Default 1.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

Value

Input object with optimized factorization values updated. including the W matrix in liger object, and W and V matrices in each ligerDataset object in the datasets slot.

Examples

pbmc <- normalize(pbmc)
#> Sat Nov  4 14:14:40 2023 ... Normalizing dataset: ctrl
#> Sat Nov  4 14:14:40 2023 ... Normalizing dataset: stim
pbmc <- selectGenes(pbmc)
#> Sat Nov  4 14:14:40 2023 ... Selecting variable features for dataset: ctrl
#> Sat Nov  4 14:14:40 2023 ...... 168 features selected out of 249 shared features
#> Sat Nov  4 14:14:40 2023 ... Selecting variable features for dataset: stim
#> Sat Nov  4 14:14:40 2023 ...... 166 features selected out of 249 shared features
#> Sat Nov  4 14:14:40 2023 ... Finally 173 shared variable features selected.
pbmc <- scaleNotCenter(pbmc)
#> Sat Nov  4 14:14:40 2023 ... Scaling dataset: ctrl
#> Sat Nov  4 14:14:40 2023 ... Scaling dataset: stim
# Only running a few iterations for fast examples
pbmc <- optimizeALS(pbmc, k = 20, maxIter = 2)
pbmc <- optimizeNewLambda(pbmc, new = 5.5, max.iters = 2)