Perform factorization for subset of data

Uses an efficient strategy for updating that takes advantage of the information in the existing factorization.

optimizeSubset(
  object,
  cellIdx,
  lambda = NULL,
  thresh = 1e-04,
  max.iters = 100,
  datasets.scale = NULL,
  seed = 1,
  verbose = getOption("ligerVerbose"),
  cell.subset = NULL,
  cluster.subset = NULL
)

Arguments

object

liger object. Should call optimizeALS in advance.

cellIdx

Valid index vector that applies to the whole object. See subsetLiger for requirement.

lambda

Numeric regularization parameter. Default 5.

thresh

Convergence threshold. Convergence occurs when \(|obj_0-obj|/(mean(obj_0,obj)) < thresh\). Default 1e-4.

max.iters

Maximum number of block coordinate descent iterations to perform. Default 100.

datasets.scale

Names of datasets to rescale after subsetting. Default NULL does not rescale.

seed

Random seed to allow reproducible results. Default 1.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

cell.subset, cluster.subset

Deprecated. Please use cellIdx to explicitly specify.

Value

Subset object with factorization matrices reset, including the W matrix in liger object, and W and V

matrices in each ligerDataset object in the datasets

slot. scaleData in the ligerDataset objects of datasets specified by datasets.scale will also be updated to reflect the subset.

Examples

pbmc <- normalize(pbmc)
#> Mon Nov  6 16:33:50 2023 ... Normalizing dataset: ctrl
#> Mon Nov  6 16:33:50 2023 ... Normalizing dataset: stim
pbmc <- selectGenes(pbmc)
#> Mon Nov  6 16:33:50 2023 ... Selecting variable features for dataset: ctrl
#> Mon Nov  6 16:33:50 2023 ...... 171 features selected out of 249 shared features
#> Mon Nov  6 16:33:50 2023 ... Selecting variable features for dataset: stim
#> Mon Nov  6 16:33:50 2023 ...... 249 features selected out of 249 shared features
#> Mon Nov  6 16:33:50 2023 ... Finally 171 shared variable features selected.
pbmc <- scaleNotCenter(pbmc)
#> Mon Nov  6 16:33:50 2023 ... Scaling dataset: ctrl
#> Mon Nov  6 16:33:50 2023 ... Scaling dataset: stim
# Only running a few iterations for fast examples
pbmc <- optimizeALS(pbmc, k = 20, maxIter = 2)
pbmc <- optimizeSubset(pbmc, cellIdx = sort(sample(ncol(pbmc), 200)),
                       max.iters = 2)
#> Mon Nov  6 16:33:50 2023 ... Subsetting dataset: ctrl
#> Mon Nov  6 16:33:50 2023 ... Subsetting dataset: stim
#> Error: Each given H must be of size Ei[i].n_cols x 20