Generate Dimensionality Reduction Plot with Coloring

some text

plotClusterDimRed(object, useCluster = NULL, useDimRed = NULL, ...)

plotDatasetDimRed(object, useDimRed = NULL, ...)

plotByDatasetAndCluster(
  object,
  useDimRed = NULL,
  useCluster = NULL,
  combinePlots = TRUE,
  ...
)

plotGeneDimRed(
  object,
  features,
  useDimRed = NULL,
  log = TRUE,
  scaleFactor = 10000,
  zeroAsNA = TRUE,
  colorPalette = "C",
  ...
)

plotPeakDimRed(
  object,
  features,
  useDimRed = NULL,
  log = TRUE,
  scaleFactor = 10000,
  zeroAsNA = TRUE,
  colorPalette = "C",
  ...
)

plotFactorDimRed(
  object,
  factors,
  useDimRed = NULL,
  trimHigh = 0.03,
  zeroAsNA = TRUE,
  colorPalette = "D",
  ...
)

Arguments

object

A liger object.

useCluster

Name of variable in cellMeta slot. Default NULL uses default cluster.

useDimRed

Name of the variable storing dimensionality reduction result in the cellMeta slot. Default NULL use default dimRed.

...

Additional graphic setting arguments passed to plotCellScatter.

combinePlots

Logical, whether to utilize plot_grid to combine multiple plots into one. Default TRUE returns combined ggplot. FALSE returns a list of ggplot.

features, factors

Name of genes or index of factors that need to be visualized.

log

Logical. Whether to log transform the normalized expression of genes. Default TRUE.

scaleFactor

Number to be multiplied with the normalized expression of genes before log transformation. Default 1e4. NULL for not scaling.

zeroAsNA

Logical, whether to swap all zero values to NA so naColor will be used to represent non-expressing features. Default TRUE.

colorPalette

Name of viridis palette. See viridis for options. Default "C" ("plasma") for gene expression and "D" ("viridis") for factor loading.

trimHigh

Number for highest cut-off to limit the outliers. Factor loading above this value will all be trimmed to this value. Default 0.03.

Value

ggplot object when only one feature (e.g. cluster variable, gene, factor) is set. List object when multiple of those are specified.

See also

Examples

plotClusterDimRed(pbmcPlot)
#> Error in plotClusterDimRed(pbmcPlot): No `useDimRed` given or default dimRed not set.
plotDatasetDimRed(pbmcPlot)
#> Error in plotDatasetDimRed(pbmcPlot): No `useDimRed` given or default dimRed not set.
plotByDatasetAndCluster(pbmcPlot)
#> Error in plotDatasetDimRed(object, useDimRed = useDimRed, ...): No `useDimRed` given or default dimRed not set.
plotGeneDimRed(pbmcPlot, varFeatures(pbmcPlot)[1])
#> Warning: Specified variables from cellMeta not found: .1, .2
#> Sat Nov  4 14:14:47 2023 ... Plotting feature: ISG15, on 600 cells...
#> Error in plotDF[[colorBy]]: subscript out of bounds
plotFactorDimRed(pbmcPlot, 2)
#> Warning: Specified variables from cellMeta not found: .1, .2
#> Sat Nov  4 14:14:47 2023 ... Plotting feature: Factor_19, on 600 cells...
#> Error in plotDF[[colorBy]]: subscript out of bounds