Restore links (to HDF5 files) for reloaded liger/ligerDataset object

When loading the saved liger object with HDF5 data in a new R session, the links to HDF5 files would be corrupted. This functions enables the restoration of those links so that new analyses can be carried out.

restoreH5Liger(object, filePath = NULL)

restoreOnlineLiger(object, file.path = NULL)

Arguments

object

liger or ligerDataset object.

filePath

Paths to HDF5 files. A single character path for ligerDataset input or a list of paths named by the datasets for liger object input. Default NULL looks for the path(s) of the last valid loading.

file.path

Will be deprecated with restoreOnlineLiger. The same as filePath.

Value

object with restored links.

Note

restoreOnlineLiger will be deprecated for clarifying the terms used for data structure.