Doublet detection with DoubletFinder

Detect doublet with DoubletFinder. Package "Seurat" and "DoubletFinder" would be required to run this function. "DoubletFinder" can only be installed via GitHub.

Seurat default PCA workflow is performed prior to calling DoubletFinder.

runDoubletFinder(
  object,
  useDatasets = NULL,
  nPCs = 10,
  nNeighbors = 20,
  nExp = NULL,
  verbose = getOption("ligerVerbose"),
  ...
)

Arguments

object

A liger object.

useDatasets

A character vector of the names, a numeric or logical vector of the index of the datasets to run doubletFinder_v3 with. Default NULL applies to all datasets.

nPCs

Number of top PC to use. Default 10.

nNeighbors

Number of the PC neighborhood size used to compute pANN. See "See Also". Scalar for all used datasets or vector for each. Default 20.

nExp

The total number of doublet predictions produced. Scalar for all used datasets or vector for each. Default NULL sets a 0.15 proportion.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

...

Additional arguments passed to doubletFinder_v3.

Value

Updated object with variables DoubletFinder_pANN and DoubletFinder_classification updated in cellMeta slot

See also

Examples

pbmc <- runDoubletFinder(pbmc)
#> Error in `cellMeta<-`(`*tmp*`, "DoubletFinder_pANN", value = NA): Length of value does not match with cell index.
cellMeta(pbmc)
#> DataFrame with 600 rows and 7 columns
#>                        dataset               barcode      nUMI     nGene
#>                       <factor>           <character> <numeric> <integer>
#> ctrl_AAACATACCTCGCT.1     ctrl ctrl_AAACATACCTCGCT.1      2151       102
#> ctrl_AAACGGCTCTTCGC.1     ctrl ctrl_AAACGGCTCTTCGC.1      1916       103
#> ctrl_AACACTCTAAGTAG.1     ctrl ctrl_AACACTCTAAGTAG.1      1869        95
#> ctrl_AACCGCCTCAGGAG.1     ctrl ctrl_AACCGCCTCAGGAG.1       733        84
#> ctrl_AACGTTCTTCCGTC.1     ctrl ctrl_AACGTTCTTCCGTC.1       573        76
#> ...                        ...                   ...       ...       ...
#> stim_TTCATGACTTATCC.1     stim stim_TTCATGACTTATCC.1      1091       101
#> stim_TTCATGACTTCAGG.1     stim stim_TTCATGACTTCAGG.1       431        86
#> stim_TTCGGAGATTTCAC.1     stim stim_TTCGGAGATTTCAC.1      1344       102
#> stim_TTGACACTTCCTGC.1     stim stim_TTGACACTTCCTGC.1      1423       105
#> stim_TTTGCATGAACGAA.1     stim stim_TTTGCATGAACGAA.1      2818       120
#>                            mito      ribo      hemo
#>                       <numeric> <numeric> <numeric>
#> ctrl_AAACATACCTCGCT.1         0  10.83217         0
#> ctrl_AAACGGCTCTTCGC.1         0  20.25052         0
#> ctrl_AACACTCTAAGTAG.1         0   5.08293         0
#> ctrl_AACCGCCTCAGGAG.1         0  42.70123         0
#> ctrl_AACGTTCTTCCGTC.1         0  38.04538         0
#> ...                         ...       ...       ...
#> stim_TTCATGACTTATCC.1         0   8.61595         0
#> stim_TTCATGACTTCAGG.1         0  41.29930         0
#> stim_TTCGGAGATTTCAC.1         0   8.85417         0
#> stim_TTGACACTTCCTGC.1         0  11.94659         0
#> stim_TTTGCATGAACGAA.1         0   9.43932         0