Run Gene Ontology enrichment analysis on metagenes

This function forms genesets basing on the differential expression result, and calls gene ontology (GO) analysis method provided by gprofiler2.

runGOEnrich(
  result,
  group = NULL,
  useBg = TRUE,
  orderBy = "logFC",
  logFCThresh = 1,
  padjThresh = 0.01,
  ...
)

Arguments

result

Data frame of unfiltered output from runWilcoxon.

group

Selection of one group available from result$group. Default NULL uses all groups involved in DE result table.

useBg

Logical, whether to set all genes involved in DE analysis (before threshold filtering) as a domain background of GO analysis. Default TRUE.

orderBy

Name of DE statistics metric to order the gene list for each group. Choose from "logFC" (default), "pval" or "padj". Or set NULL to turn off ranked mode.

logFCThresh

The log2FC threshold above which the genes will be used. Default 1.

padjThresh

The adjusted p-value threshold less than which the genes will be used. Default 0.01.

...

Additional arguments passed to gost.

Value

A list object with the following entries

result

data.frame of main GO analysis result.

meta

Meta information for the query.

See gost. for detailed explanation.

References

Kolberg, L. et al, 2020 and Raudvere, U. et al, 2019

Examples

res <- runWilcoxon(pbmcPlot)
#> Mon Nov  6 16:34:08 2023 ... Performing Wilcoxon test on 2 datasets: ctrl, stim
# Setting `significant = FALSE` because it's hard for a gene list obtained
# from small test dataset to represent real-life biology.
go <- runGOEnrich(res, group = 0, significant = FALSE)