Analyze biological interpretations of metagene

Identify the biological pathways (gene sets from Reactome) that each metagene (factor) might belongs to.

runGSEA(
  object,
  genesets = NULL,
  useW = TRUE,
  useDatasets = NULL,
  customGenesets = NULL,
  gene_sets = genesets,
  mat_w = useW,
  mat_v = useDatasets,
  custom_gene_sets = customGenesets
)

Arguments

object

A liger object with valid factorization result.

genesets

Character vector of the Reactome gene sets names to be tested. Default NULL uses all the gene sets from the Reactome.

useW

Logical, whether to use the shared factor loadings (\(W\)). Default TRUE.

useDatasets

A character vector of the names, a numeric or logical vector of the index of the datasets where the \(V\) matrices will be included for analysis. Default NULL uses all datasets.

customGenesets

A named list of character vectors of entrez gene ids. Default NULL uses all the gene symbols from the input matrix.

gene_sets, mat_w, mat_v, custom_gene_sets

Deprecated. See Usage section for replacement.

Value

A list of matrices with GSEA analysis for each factor

Examples

runGSEA(pbmcPlot)
#> 
#> 'select()' returned 1:1 mapping between keys and columns
#> 'select()' returned 1:many mapping between keys and columns
#> 'select()' returned 1:1 mapping between keys and columns
#> Warning: You are trying to run fgseaSimple. It is recommended to use fgseaMultilevel. To run fgseaMultilevel, you need to remove the nperm argument in the fgsea function call.
#> Warning: There are ties in the preranked stats (34.04% of the list).
#> The order of those tied genes will be arbitrary, which may produce unexpected results.
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> Warning: You are trying to run fgseaSimple. It is recommended to use fgseaMultilevel. To run fgseaMultilevel, you need to remove the nperm argument in the fgsea function call.
#> Warning: There are ties in the preranked stats (40.43% of the list).
#> The order of those tied genes will be arbitrary, which may produce unexpected results.
#> Warning: All values in the stats vector are greater than zero and scoreType is "std", maybe you should switch to scoreType = "pos".
#> $Factor_1
#>                    pathway      pval       padj        ES      NES nMoreExtreme
#> 1            Immune System 0.0126327 0.05053079 0.6137536 1.637542          118
#> 2                  Disease 0.2718995 0.27189953 0.4293680 1.172129         2597
#> 3       Infectious disease 0.2074755 0.27189953 0.4570513 1.228547         1964
#> 4 Viral Infection Pathways 0.1865290 0.27189953 0.4719430 1.250740         1752
#>   size  leadingEdge
#> 1   17 91543, 9....
#> 2   21 9636, 92....
#> 3   18 9636, 61....
#> 4   16 9636, 61....
#> 
#> $Factor_2
#>                    pathway         pval         padj         ES       NES
#> 1                  Disease 0.0001051967 0.0001417384  0.7911480  2.166210
#> 2       Infectious disease 0.0001056078 0.0001417384  0.8289855  2.248346
#> 3 Viral Infection Pathways 0.0001063038 0.0001417384  0.8613096  2.309901
#> 4            Immune System 0.1652323580 0.1652323580 -0.3033898 -1.240585
#>   nMoreExtreme size  leadingEdge
#> 1            0   21 6164, 61....
#> 2            0   18 6164, 61....
#> 3            0   16 6164, 61....
#> 4           95   17 929, 362....
#>