Integrate scaled datasets with iNMF or variant methods

LIGER provides dataset integration methods based on iNMF (integrative Non-negative Matrix Factorization [1]) and its variants (online iNMF [2] and UINMF [3]). This function wraps runINMF, runOnlineINMF and runUINMF, of which the help pages have more detailed description.

runIntegration(
  object,
  k = 20,
  lambda = 5,
  method = c("iNMF", "onlineINMF", "UINMF"),
  seed = 1,
  verbose = getOption("ligerVerbose"),
  ...
)

# S3 method for liger
runIntegration(
  object,
  k = 20,
  lambda = 5,
  method = c("iNMF", "onlineINMF", "UINMF"),
  seed = 1,
  verbose = getOption("ligerVerbose"),
  ...
)

# S3 method for Seurat
runIntegration(
  object,
  k = 20,
  lambda = 5,
  method = c("iNMF", "onlineINMF"),
  datasetVar = "orig.ident",
  useLayer = "ligerScaleData",
  assay = NULL,
  seed = 1,
  verbose = getOption("ligerVerbose"),
  ...
)

Arguments

object

A liger object or a Seurat object with non-negative scaled data of variable features (Done with scaleNotCenter).

k

Inner dimension of factorization (number of factors). Generally, a higher k will be needed for datasets with more sub-structure. Default 20.

lambda

Regularization parameter. Larger values penalize dataset-specific effects more strongly (i.e. alignment should increase as lambda increases). Default 5.

seed

Random seed to allow reproducible results. Default 1.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

...

Arguments passed to other methods and wrapped functions.

datasetVar

Metadata variable name that stores the dataset source annotation. Default "orig.ident".

useLayer

For Seurat>=4.9.9, the name of layer to retrieve input non-negative scaled data. Default "ligerScaleData". For older Seurat, always retrieve from scale.data slot.

assay

Name of assay to use. Default NULL uses current active assay.

Value

Updated input object. For detail, please refer to the refered method linked in Description.

References

Joshua D. Welch and et al., Single-Cell Multi-omic Integration Compares and Contrasts Features of Brain Cell Identity, Cell, 2019

Chao Gao and et al., Iterative single-cell multi-omic integration using online learning, Nat Biotechnol., 2021

April R. Kriebel and Joshua D. Welch, UINMF performs mosaic integration of single-cell multi-omic datasets using nonnegative matrix factorization, Nat. Comm., 2022

Examples

pbmc <- normalize(pbmc)
#> Mon Dec 11 16:11:30 2023 ... Normalizing dataset: ctrl
#> Mon Dec 11 16:11:30 2023 ... Normalizing dataset: stim
pbmc <- selectGenes(pbmc)
#> Mon Dec 11 16:11:30 2023 ... Selecting variable features for dataset: ctrl
#> Mon Dec 11 16:11:30 2023 ...... 168 features selected out of 249 shared features
#> Mon Dec 11 16:11:30 2023 ... Selecting variable features for dataset: stim
#> Mon Dec 11 16:11:30 2023 ...... 166 features selected out of 249 shared features
#> Mon Dec 11 16:11:30 2023 ... Finally 173 shared variable features selected.
pbmc <- scaleNotCenter(pbmc)
#> Mon Dec 11 16:11:30 2023 ... Scaling dataset: ctrl
#> Mon Dec 11 16:11:30 2023 ... Scaling dataset: stim
pbmc <- runIntegration(pbmc)