Run DESeq2 differential expression test on pseudo-bulk expression

This function aggregate raw counts data of specified comparison groups of cells basing on either given biological replicate annotation or pseudo replicates. According to user prior knowledge, existing replicate annotation should be used when it is true that the cells involved in comparison come from multiple biological replicates, while when all cells in both groups come from the same replicate, pseudo replicates should be created.

Known replicate annotation can be specified by useCellMetaVar to use metadata stored in the cellMeta slot, or specify external annotation through replicateAnn. To use pseudo replicates, users should leave the two mentioned argument as default NULL.

runPseudoBulkDEG(
  object,
  groups = NULL,
  markerBy = NULL,
  method = c("DESeq2", "limma", "edgeR"),
  useCellMetaVar = NULL,
  replicateAnn = NULL,
  nRep = 5,
  seed = 1,
  verbose = getOption("ligerVerbose")
)

Arguments

object

A liger object.

groups

A named list with at least two elements, where each is a vector of valid cell specification (numeric or logical indexing should be based on the whole object). See Details. Default NULL.

markerBy

cellMeta categorical variable name(s) to group cells. See Details. Default NULL.

method

Choose from "DESeq2", "limma" or "edgeR". Default "DESeq2"

useCellMetaVar

Name(s) of cellMeta variable that specify the replicate annotation. Default NULL.

replicateAnn

A data.frame or factor to pass external replicate annotation. See Details for requirements. Default NULL.

nRep

Number of pseudo replicates per comparison group, applied only when running pseudo replicate mode (i.e. both useCellMetaVar and replicateAnn are NULL). Default 5.

seed

Random seed to allow reproducible results. Default 1.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

Value

A data.frame as the output of DESeq2. Ordered by adjusted p-value (padj).

Details

The differential expression test can be run in two mode:

Group wise test: Users need to explicitly specify the cell groups with a named list through groups, and leave markerBy as NULL. The DE test will be performed pairwisely between groups.

Biomarker test: Users need to pass the grouping (e.g. clustering) variable(s) to markerBy, and leave groups as NULL. The DE test will be performed by comparing one group against all the rest iteratively.

For using replicateAnn. When passing a data.frame, rownames have to be available and it should cover all cells involved in both comparison group. When using a named factor/vector, the names should as well cover all involved cells. When using a un-named factor/vector, it is recommended to have its length matched with the total number of cells of the whole object. When the length of the un-named factor/vector equals to length(groups[[1]]) + length(groups[[2]]), it is assumed that the order matches to c(groups[[1]], groups[[2]]).

Examples

rawData(datasets(pbmcPlot)[[1]]) <- rawData(dataset(pbmc, 1))
rawData(datasets(pbmcPlot)[[2]]) <- rawData(dataset(pbmc, 2))
res <- runPseudoBulkDEG(
    pbmcPlot,
    groups = list(
        c1 = pbmcPlot$leiden_cluster == 1,
        c2 = pbmcPlot$leiden_cluster == 2
    )
)
#> Sun Nov  5 11:38:14 2023 ... No replicate annotation specified. Creating pseudo-replicates...
#> Sun Nov  5 11:38:14 2023 ... Subsetting dataset: ctrl
#> Sun Nov  5 11:38:14 2023 ... Subsetting dataset: stim
#> Sun Nov  5 11:38:14 2023 ... Generating pseudo-bulks for condition "c1":
#>                              c2.6, c2.7, c2.8, c2.9, c2.10
#> Sun Nov  5 11:38:14 2023 ... Generating pseudo-bulks for condition "c2":
#>                              c1.1, c1.2, c1.3, c1.4, c1.5
#> Sun Nov  5 11:38:14 2023 ... Calling DESeq2 Wald test
#> converting counts to integer mode
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
print(head(res))
#>    feature     logFC          pval          padj
#> 8  S100A11 -4.866363 2.272674e-122 1.136337e-120
#> 7  S100A10 -4.563204 1.766943e-119 4.417356e-118
#> 11  S100A6 -4.442083  1.157403e-69  1.929005e-68
#> 41   ANXA5 -6.564061  2.602630e-59  3.253288e-58
#> 12  S100A4 -6.355146  8.603406e-59  8.603406e-58
#> 14  FCER1G -7.250794  3.623027e-57  3.019189e-56