Perform t-SNE dimensionality reduction

Runs t-SNE on the quantile normalized cell factors (result from quantileNorm), or unnormalized cell factors (result from optimizeALS or online_iNMF)) to generate a 2D embedding for visualization. By default Rtsne (Barnes-Hut implementation of t-SNE) method is invoked, while alternative "fftRtsne" method (FFT-accelerated Interpolation-based t-SNE, using Kluger Lab implementation) is also supported.

In order to run fftRtsne (recommended for large datasets), FIt-SNE must be installed as instructed in detailed here. Include the path to the cloned FIt-SNE directory as the fitsne.path parameter, though this is only necessary for the first call to run runTSNE. For more detailed FIt-SNE installation instructions, see the liger repo README.

runTSNE(
  object,
  useRaw = FALSE,
  useDims = NULL,
  nDims = 2,
  usePCA = FALSE,
  perplexity = 30,
  theta = 0.5,
  method = c("Rtsne", "fftRtsne"),
  dimredName = "TSNE",
  fitsnePath = NULL,
  seed = 42,
  verbose = getOption("ligerVerbose"),
  k = nDims,
  use.raw = useRaw,
  dims.use = useDims,
  use.pca = usePCA,
  fitsne.path = fitsnePath,
  rand.seed = seed
)

Arguments

object

liger object with factorization results.

useRaw

Whether to use un-aligned cell factor loadings (\(H\) matrices). Default FALSE.

useDims

Index of factors to use for computing UMAP embedding. Default NULL uses all factors.

nDims

Number of dimensions to reduce to. Default 2.

usePCA

Whether to perform initial PCA step for Rtsne. Default FALSE.

perplexity

Numeric parameter to pass to Rtsne (expected number of neighbors). Default 30.

theta

Speed/accuracy trade-off (increase for less accuracy), set to 0.0 for exact TSNE. Default 0.5.

method

Choose from "Rtsne" or "fftRtsne". See Description. Default "Rtsne".

dimredName

Name of the variable in cellMeta slot to store the result matrix. Default "TSNE".

fitsnePath

Path to the cloned FIt-SNE directory (i.e. '/path/to/dir/FIt-SNE'). Required only when first time using runTSNE with method = "fftRtsne". Default NULL.

seed

Random seed for reproducibility. Default 42.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

use.raw, dims.use, k, use.pca, fitsne.path, rand.seed

Deprecated. See Usage section for replacement.

Value

The object where a "TSNE" variable is updated in the cellMeta slot with the whole 2D embedding matrix.

See also

Examples

pbmc <- runTSNE(pbmcPlot)
#> Sun Nov  5 17:13:44 2023 ... Generating TSNE (Rtsne) on quantile normalized cell factor loadings...