Perform Wilcoxon rank-sum test

Perform Wilcoxon rank-sum tests on specified dataset using given method.

runWilcoxon(
  object,
  useDatasets = NULL,
  method = c("clusters", "datasets"),
  useCluster = "leiden_cluster",
  usePeak = FALSE,
  verbose = getOption("ligerVerbose"),
  data.use = useDatasets,
  compare.method = method
)

Arguments

object

A liger object with cluster assignment available.

useDatasets

A character vector of the names, a numeric or logical vector of the index of the datasets to be normalized. Default NULL uses all datasets.

method

Choose from "clusters" or "datasets". Default "clusters" compares between clusters across all datasets, while "datasets" compares between datasets within each cluster.

useCluster

The name of the column in cellMeta slot storing the cluster assignment variable. Default "leiden_cluster"

usePeak

Logical, whether to test peak counts instead of gene expression. Requires presence of ATAC modility datasets. Default FALSE.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

data.use, compare.method

Deprecated. See Usage section for replacement.

Value

A 10-columns data.frame with test results.

Examples

library(dplyr)
result <- runWilcoxon(pbmcPlot)
#> Sun Nov  5 17:13:48 2023 ... Performing Wilcoxon test on 2 datasets: ctrl, stim
result %>% group_by(group) %>% top_n(2, logFC)
#> # A tibble: 14 × 10
#> # Groups:   group [7]
#>    feature  group avgExpr logFC statistic   auc     pval     padj pct_in pct_out
#>    <chr>    <fct>   <dbl> <dbl>     <dbl> <dbl>    <dbl>    <dbl>  <dbl>   <dbl>
#>  1 GNLY     0        6.76  6.18    47936. 0.665 4.77e-28 3.97e-27   35.5    3.46
#>  2 CXCR4    0       14.9   7.81    58483  0.812 1.35e-35 6.75e-34   84.3   42.2 
#>  3 S100A4   1       16.6  11.3     57948  0.887 4.83e-52 8.05e-51   94.4   31.5 
#>  4 IL8      1       14.1  11.7     55230  0.845 3.60e-54 9.01e-53   79.0   13.8 
#>  5 SELL     2       11.3   7.74    41900. 0.746 2.14e-24 7.12e-24   65.5   21.3 
#>  6 NPM1     2       13.7   6.03    40044. 0.713 3.74e-14 9.35e-14   79.3   45.5 
#>  7 RSAD2    3       16.8  11.4     36136. 0.897 2.02e-37 1.01e-35   93.5   31.4 
#>  8 CXCL10   3       16.8  12.1     34710. 0.862 2.08e-34 5.19e-33   85.7   25.6 
#>  9 ID3      4       11.0   8.92    15734. 0.775 1.53e-18 5.80e-17   63.9   12.4 
#> 10 CD83     4       13.1   9.81    16686. 0.822 2.32e-18 5.80e-17   75     20.0 
#> 11 FCGR3A   5       17.2  13.6     17106  0.970 2.16e-30 1.08e-28   96.8   22.0 
#> 12 CXCL10   5       16.0  10.3     13906. 0.788 1.12e-10 1.40e- 9   83.9   30.6 
#> 13 MARCKSL1 6       17.0  11.0     15232  0.864 1.06e-14 1.77e-13  100     36.2 
#> 14 CD83     6       15.8  12.5     15670. 0.888 8.08e-23 4.04e-21   93.5   19.5