Scale genes by root-mean-square across cells

This function scales normalized gene expression data after variable genes have been selected. We do not mean-center the data before scaling in order to address the non-negativity constraint of NMF. Computation applied to each normalized dataset matrix can form the following equation:

$$S_{i,j}=\frac{N_{i,j}}{\sqrt{\sum_{p}^{n}\frac{N_{i,p}^2}{n-1}}}$$

Where \(N\) denotes the normalized matrix for an individual dataset, \(S\) is the output scaled matrix for this dataset, and \(n\) is the number of cells in this dataset. \(i, j\) denotes the specific gene and cell index, and \(p\) is the cell iterator.

scaleNotCenter(object, ...)

# S3 method for dgCMatrix
scaleNotCenter(object, ...)

# S3 method for ligerDataset
scaleNotCenter(
  object,
  features = NULL,
  chunk = 1000,
  verbose = getOption("ligerVerbose"),
  ...
)

# S3 method for liger
scaleNotCenter(
  object,
  useDatasets = NULL,
  features = varFeatures(object),
  verbose = getOption("ligerVerbose"),
  remove.missing = NULL,
  ...
)

# S3 method for Seurat
scaleNotCenter(
  object,
  assay = NULL,
  layer = "ligerScaleData",
  useLayer = "ligerNormData",
  datasetVar = "orig.ident",
  features = NULL,
  ...
)

Arguments

object

liger object, ligerDataset object, dgCMatrix, or a Seurat object.

...

Arguments passed to other methods. The order goes by: "liger" method calls "ligerDataset" method", which then calls "dgCMatrix" method. "Seurat" method directly calls "dgCMatrix" method.

features

Character, numeric or logical index that choose the variable feature to be scaled. "liger" method by default uses varFeatures(object). "ligerDataset" method by default uses all features. "Seurat" method by default uses VariableFeatures(object).

chunk

Integer. Number of maximum number of cells in each chunk, when scaling is applied to any HDF5 based dataset. Default 1000.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

useDatasets

A character vector of the names, a numeric or logical vector of the index of the datasets to be scaled but not centered. Default NULL applies to all datasets.

remove.missing

Deprecated. The functionality of this is covered through other parts of the whole workflow and is no long needed. Will be ignored if specified.

assay

Name of assay to use. Default NULL uses current active assay.

layer

For Seurat>=4.9.9, the name of layer to store normalized data. Default "ligerScaleData". For older Seurat, stored to scale.data slot.

useLayer

For Seurat>=4.9.9, the name of layer to retrieve normalized data. Default "ligerNormData". For older Seurat, always retrieve from data slot.

datasetVar

Metadata variable name that stores the dataset source annotation. Default "orig.ident".

Value

Updated object

dgCMatrix method - Returns scaled dgCMatrix object

ligerDataset method - Updates the scaleData and scaledUnsharedData (if unshared variable feature available) slot of the object

liger method - Updates the scaleData and scaledUnsharedData (if unshared variable feature available) slot of chosen datasets

Seurat method - Adds a named layer in chosen assay (V5), or update the scale.data slot of the chosen assay (<=V4)

Note

Since the scaling on genes is applied on a per dataset base, other scaling methods that apply to a whole concatenated matrix of multiple datasets might not be considered as equivalent alternatives, even if options like center are set to FALSE. Hence we implemented an efficient solution that works under such circumstance, provided with the Seurat S3 method.

Examples

pbmc <- normalize(pbmc)
#> Wed Dec 20 15:51:16 2023 ... Normalizing dataset: ctrl
#> Wed Dec 20 15:51:16 2023 ... Normalizing dataset: stim
pbmc <- selectGenes(pbmc)
#> Wed Dec 20 15:51:16 2023 ... Selecting variable features for dataset: ctrl
#> Wed Dec 20 15:51:16 2023 ...... 168 features selected out of 249 shared features
#> Wed Dec 20 15:51:16 2023 ... Selecting variable features for dataset: stim
#> Wed Dec 20 15:51:16 2023 ...... 166 features selected out of 249 shared features
#> Wed Dec 20 15:51:16 2023 ... Finally 173 shared variable features selected.
pbmc <- scaleNotCenter(pbmc)
#> Wed Dec 20 15:51:16 2023 ... Scaling dataset: ctrl
#> Wed Dec 20 15:51:16 2023 ... Scaling dataset: stim