Select variable gene from one dataset with VST method

A re-implimentation of Seurat FindVariableFeatures VST method. This allows the selection of a fixed number of variable features from one dataset.

selectGenesVST(
  object,
  useDataset,
  n = 2000,
  loessSpan = 0.3,
  clipMax = "auto",
  verbose = getOption("ligerVerbose")
)

Arguments

object

A liger object

useDataset

The names, a numeric or logical index of the ONE dataset to be considered for selection.

n

Number of variable features needed. Default 2000.

loessSpan

Loess span parameter used when fitting the variance-mean relationship. Default 0.3.

clipMax

After standardization values larger than clipMax will be set to clipMax. Default "auto" sets this value to the square root of the number of cells.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

References

Seurat::FindVariableFeatures.default(selection.method = "vst")

Examples

pbmc <- selectGenesVST(pbmc, "ctrl", n = 50)
#> Sun Nov  5 11:38:20 2023 ... Selecting top 50 HVGs with VST method for dataset: ctrl