Visually suggest appropriate lambda value

Can be used to select appropriate value of lambda for factorization of particular dataset. Plot alignment and agreement for various test values of lambda. Most appropriate lambda is likely around the "elbow" of the alignment plot (when alignment stops increasing). This will likely also correspond to slower decrease in agreement. Depending on number of cores used, this process can take 10-20 minutes.

suggestLambda(
  object,
  k,
  lambdaTest = NULL,
  seed = 1,
  nCores = 1,
  thresh = 1e-04,
  maxIter = 100,
  nNeighbors = 20,
  reference = NULL,
  genNew = FALSE,
  nrep = 1,
  returnRaw = FALSE,
  verbose = getOption("ligerVerbose"),
  lambda.test = lambdaTest,
  rand.seed = seed,
  num.cores = nCores,
  max.iters = maxIter,
  knn_k = nNeighbors,
  ref_dataset = reference,
  gen.new = genNew,
  return.raw = returnRaw,
  return.data = TRUE,
  k2 = 500,
  resolution = 1
)

Arguments

object

A liger object with scaled data.

k

Number of factors to use in the test.

lambdaTest

Numeric vector of "lambda" values to test. Default NULL will test the following sequence: 0.25, 0.5, 0.75; 1, 2 to 10; and 15, 20, ... to 60, totally 23 numbers.

seed

Random seed to allow reproducible results. Default 1.

nCores

Number of cores to use for optimizing factorization in parallel. Default 1.

thresh

Convergence threshold. Convergence occurs when \(|obj_0-obj|/(mean(obj_0,obj)) < thresh\). Default 1e-4.

maxIter

Maximum number of block coordinate descent iterations to perform. Default 100.

nNeighbors

Number of nearest neighbors for within-dataset KNN in quantileNorm. Default 20.

reference

Reference dataset for quantileNorm. See link for detail. Default NULL.

genNew

Logical. Whether to use optimizeNewLambda in factorization. Recommended TRUE when looking at only a small range of lambdas (i.e. 1:7). Default FALSE.

nrep

See optimizeALS for detail. Number restarts to perform at each k value tested. Increase to produce smoother curve if results unclear. Default 1.

returnRaw

Logical. Whether to return raw data or data.frame. Raw data is matrix of alignment values for each lambda value tested (each column represents a different rep for nrep). Default FALSE.

verbose

Logical. Whether to show information of the progress. Default getOption("ligerVerbose") which is TRUE if users have not set.

lambda.test, rand.seed, num.cores, max.iters, knn_k, ref_dataset, gen.new, return.raw

Deprecated. See Usage section for replacement.

return.data

Defuncted. Will always return data now. Use plotSuggestLambda to show the plot.

k2, resolution

Defuncted. Parameter passed to quantileAlignSNF which is deprecated.

Value

Matrix of results